implémentation des metric ETC,ITC,NOA,NOC,NOD,NOM et CAC
Les métric ETC,ITC,NOA,NOC,NOD,NOM sont fonctonnel.
Le metric CAC ne fonctionne pas parfaitement il doit être ajusté

+++ tp2/src/metrics/MetricNOA.java
@@ -21,7 +21,7 @@ public class MetricNOA extends BaseMetric {
 		
 		List<Attribute> listAttribute=c.getAttributes();
 		List<UMLClass> listClass= m.getClasses();
-		float NOA= listAttribute.size();
+		double NOA= listAttribute.size();
 		
 			for(int j=0;j<listClass.size();j++)
 			{
